requirements.txt
# Core AI
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Environment & config
python-dotenv>=1.0.0

# HTTP & async
httpx>=0.27.0
aiohttp>=3.10.0

# Data handling
pydantic>=2.0.0

# Optional agent tools (uncomment as needed)
# langchain-experimental>=0.3.0
# duckduckgo-search>=6.0.0
# beautifulsoup4>=4.12.0
# requests>=2.32.0
# mcp>=1.0.0

# Dev/debugging
rich>=13.0.0
